Now downloads data from all years 1996-2016
Had to create a separate loop for 2016 because the data is stored in 2
separate tables due to the conference system.
I had to add lots of exceptions from 2011 onwards as the wikipedia
tables had different rows these years. I left these out for now, so all
data has the same columns as 1996.